Remove PowerMock.
