+++ src/test/java/de/wasenweg/alfred/unit/ApiMetaDataReaderTest.java
@@ -11,7 +11,6 @@ import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.powermock.reflect.Whitebox;
 import org.springframework.test.context.ContextConfiguration;
 
 import java.util.List;
@@ -24,85 +23,10 @@ import static org.mockito.Mockito.when;
 public class ApiMetaDataReaderTest {
 
   @InjectMocks
-  private ApiMetaDataService apiMetaDataReader;
+  private ApiMetaDataService apiMetaDataService;
 
   @Mock
   private ComicVineService comicVineService;
-
-  @Test
-  public void setPathPartsMatchesPath() throws Exception {
-    final Comic comic = new Comic();
-    comic.setPath("/foo/DC Comics/Batman (1940)/Batman 400 (1940) bar.cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
-    assertThat(comic.getSeries()).isEqualTo("Batman");
-    assertThat(comic.getVolume()).isEqualTo("1940");
-    assertThat(comic.getNumber()).isEqualTo("400");
-    assertThat(comic.getPosition()).isEqualTo("0400.0");
-  }
-
-  @Test
-  public void setPathPartsMatchesWithoutSuffix() throws Exception {
-    final Comic comic = new Comic();
-    comic.setPath("/home/foo/src/alfred/src/test/resources/fixtures/incomplete/DC Comics/Batman (1940)/Batman 701 (1940).cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
-    assertThat(comic.getSeries()).isEqualTo("Batman");
-    assertThat(comic.getVolume()).isEqualTo("1940");
-    assertThat(comic.getPosition()).isEqualTo("0701.0");
-  }
-
-  @Test
-  public void setPathPartsMatchesFractionalIssueNumbers() throws Exception {
-    final Comic comic = new Comic();
-    comic.setPath("/foo/DC Comics/Batman (1940)/Batman 102a (1940) bar.cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
-    assertThat(comic.getSeries()).isEqualTo("Batman");
-    assertThat(comic.getVolume()).isEqualTo("1940");
-    assertThat(comic.getNumber()).isEqualTo("102a");
-    assertThat(comic.getPosition()).isEqualTo("0102.5");
-
-    comic.setPath("/foo/DC Comics/Batman (1940)/Batman 1/2 (1940) bar.cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
-    assertThat(comic.getSeries()).isEqualTo("Batman");
-    assertThat(comic.getVolume()).isEqualTo("1940");
-    assertThat(comic.getNumber()).isEqualTo("1/2");
-    assertThat(comic.getPosition()).isEqualTo("0000.5");
-
-    comic.setPath("/foo/DC Comics/Batman (1940)/Batman ½ (1940) bar.cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
-    assertThat(comic.getSeries()).isEqualTo("Batman");
-    assertThat(comic.getVolume()).isEqualTo("1940");
-    assertThat(comic.getNumber()).isEqualTo("½");
-    assertThat(comic.getPosition()).isEqualTo("0000.5");
-  }
-
-  @Test
-  public void setPathPartsRejectsNonMatchingVolumes() throws Exception {
-    final Comic comic = new Comic();
-    comic.setPath("/foo/DC Comics/Batman (2001)/Batman 400 (1940) bar.cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isNull();
-    assertThat(comic.getSeries()).isNull();
-    assertThat(comic.getVolume()).isNull();
-    assertThat(comic.getNumber()).isNull();
-    assertThat(comic.getPosition()).isNull();
-  }
-
-  @Test
-  public void setPathPartsRejectsNonMatchingSeries() throws Exception {
-    final Comic comic = new Comic();
-    comic.setPath("/foo/DC Comics/Batman (1940)/Batwoman 400 (1940) bar.cbz");
-    Whitebox.invokeMethod(this.apiMetaDataReader, "setPathParts", comic);
-    assertThat(comic.getPublisher()).isNull();
-    assertThat(comic.getSeries()).isNull();
-    assertThat(comic.getVolume()).isNull();
-    assertThat(comic.getNumber()).isNull();
-    assertThat(comic.getPosition()).isNull();
-  }
 
   @Test
   public void findVolume() throws Exception {
@@ -111,7 +35,7 @@ public class ApiMetaDataReaderTest {
     comic.setSeries("Batgirl");
     comic.setPublisher("DC Comics");
     comic.setVolume("2011");
-    final String result = Whitebox.invokeMethod(this.apiMetaDataReader, "findVolumeId",
+    final String result = this.apiMetaDataService.findVolumeId(
         comic.getPublisher(), comic.getSeries(), comic.getVolume());
     assertThat(result).isEqualTo("42604");
   }
@@ -126,13 +50,13 @@ public class ApiMetaDataReaderTest {
     comic.setSeries("Batman");
     comic.setPublisher("Carlsen Comics");
     comic.setVolume("1991");
-    String result = Whitebox.invokeMethod(this.apiMetaDataReader, "findVolumeId",
+    String result = this.apiMetaDataService.findVolumeId(
         comic.getPublisher(), comic.getSeries(), comic.getVolume());
     assertThat(result).isEqualTo("117375");
 
     comic.setPublisher("DC Comics");
     comic.setVolume("2011");
-    result = Whitebox.invokeMethod(this.apiMetaDataReader, "findVolumeId",
+    result = this.apiMetaDataService.findVolumeId(
         comic.getPublisher(), comic.getSeries(), comic.getVolume());
     assertThat(result).isEqualTo("42721");
   }
@@ -147,7 +71,7 @@ public class ApiMetaDataReaderTest {
     when(this.comicVineService.findIssuesInVolume("796", 5)).thenReturn(TestHelper.parseJson("issues-batman-page6.json"));
     when(this.comicVineService.findIssuesInVolume("796", 6)).thenReturn(TestHelper.parseJson("issues-batman-page7.json"));
     when(this.comicVineService.findIssuesInVolume("796", 7)).thenReturn(TestHelper.parseJson("issues-batman-page8.json"));
-    final List<JsonNode> result = Whitebox.invokeMethod(this.apiMetaDataReader, "findVolumeIssues", "796");
+    final List<JsonNode> result = this.apiMetaDataService.findVolumeIssues("796");
     assertThat(result.size()).isEqualTo(716);
   }
 
@@ -162,7 +86,7 @@ public class ApiMetaDataReaderTest {
     comic.setVolume("1940");
     comic.setNumber("701");
 
-    Whitebox.invokeMethod(this.apiMetaDataReader, "applyIssueDetails", "https://comicvine.gamespot.com/api/issue/4000-224555/", comic);
+    this.apiMetaDataService.applyIssueDetails("https://comicvine.gamespot.com/api/issue/4000-224555/", comic);
     assertThat(comic.getTitle()).isEqualTo("R.I.P. The Missing Chapter, Part 1: The Hole In Things");
     assertThat(comic.getSummary().length()).isEqualTo(391);
     assertThat(comic.getYear()).isEqualTo((short) 2010);
