+++ src/test/java/de/wasenweg/alfred/unit/ComicTest.java
@@ -1,0 +1,84 @@
+package de.wasenweg.alfred.unit;
+
+import de.wasenweg.alfred.comics.Comic;
+import org.junit.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ComicTest {
+
+  @Test
+  public void setPathPartsMatchesPath() throws Exception {
+    final Comic comic = new Comic();
+    comic.setPath("/foo/DC Comics/Batman (1940)/Batman 400 (1940) bar.cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
+    assertThat(comic.getSeries()).isEqualTo("Batman");
+    assertThat(comic.getVolume()).isEqualTo("1940");
+    assertThat(comic.getNumber()).isEqualTo("400");
+    assertThat(comic.getPosition()).isEqualTo("0400.0");
+  }
+
+  @Test
+  public void setPathPartsMatchesWithoutSuffix() throws Exception {
+    final Comic comic = new Comic();
+    comic.setPath("/home/foo/src/alfred/src/test/resources/fixtures/incomplete/DC Comics/Batman (1940)/Batman 701 (1940).cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
+    assertThat(comic.getSeries()).isEqualTo("Batman");
+    assertThat(comic.getVolume()).isEqualTo("1940");
+    assertThat(comic.getPosition()).isEqualTo("0701.0");
+  }
+
+  @Test
+  public void setPathPartsMatchesFractionalIssueNumbers() throws Exception {
+    final Comic comic = new Comic();
+    comic.setPath("/foo/DC Comics/Batman (1940)/Batman 102a (1940) bar.cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
+    assertThat(comic.getSeries()).isEqualTo("Batman");
+    assertThat(comic.getVolume()).isEqualTo("1940");
+    assertThat(comic.getNumber()).isEqualTo("102a");
+    assertThat(comic.getPosition()).isEqualTo("0102.5");
+
+    comic.setPath("/foo/DC Comics/Batman (1940)/Batman 1/2 (1940) bar.cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
+    assertThat(comic.getSeries()).isEqualTo("Batman");
+    assertThat(comic.getVolume()).isEqualTo("1940");
+    assertThat(comic.getNumber()).isEqualTo("1/2");
+    assertThat(comic.getPosition()).isEqualTo("0000.5");
+
+    comic.setPath("/foo/DC Comics/Batman (1940)/Batman ½ (1940) bar.cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isEqualTo("DC Comics");
+    assertThat(comic.getSeries()).isEqualTo("Batman");
+    assertThat(comic.getVolume()).isEqualTo("1940");
+    assertThat(comic.getNumber()).isEqualTo("½");
+    assertThat(comic.getPosition()).isEqualTo("0000.5");
+  }
+
+  @Test
+  public void setPathPartsRejectsNonMatchingVolumes() throws Exception {
+    final Comic comic = new Comic();
+    comic.setPath("/foo/DC Comics/Batman (2001)/Batman 400 (1940) bar.cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isNull();
+    assertThat(comic.getSeries()).isNull();
+    assertThat(comic.getVolume()).isNull();
+    assertThat(comic.getNumber()).isNull();
+    assertThat(comic.getPosition()).isNull();
+  }
+
+  @Test
+  public void setPathPartsRejectsNonMatchingSeries() throws Exception {
+    final Comic comic = new Comic();
+    comic.setPath("/foo/DC Comics/Batman (1940)/Batwoman 400 (1940) bar.cbz");
+    comic.setPathParts();
+    assertThat(comic.getPublisher()).isNull();
+    assertThat(comic.getSeries()).isNull();
+    assertThat(comic.getVolume()).isNull();
+    assertThat(comic.getNumber()).isNull();
+    assertThat(comic.getPosition()).isNull();
+  }
+}
